apps/groups/serializers.py
from rest_framework import serializers

from . import models


class CreateGroupSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Group
        fields = ('id', 'name')


class GetUpdateGroupSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Group
        fields = ('id', 'name', 'creator', 'created')
